fix: move react into component folder and change folder name for better folder structure
