+++ src/layouts/page.astro
@@ -1,5 +1,5 @@
 ---
-import { Navigation } from "@/components/navigation";
+import { Navigation } from "@/components/Navigation";
 import "@/styles/globals.css";
 
 const { title } = Astro.props;
